+++ reverse-sandbox/reduced_edge.py
